% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airc_id_from_path.R
\name{get_airc_id_from_path}
\alias{get_airc_id_from_path}
\title{get_airc_id_from_path}
\usage{
get_airc_id_from_path(path)
}
\arguments{
\item{path}{file path}
}
\value{
airc id (character)
}
\description{
get_airc_id_from_path
}
\examples{
path <- "/fake/dir/3tb9999/3tb9999_fake_file.txt"
get_airc_id_from_path(path)
}
\concept{helper}
