% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives_table.R
\name{get_descriptives_table}
\alias{get_descriptives_table}
\title{get_descriptives_table}
\usage{
get_descriptives_table(
  data,
  grouping_variable,
  grouping_values,
  variables = c("n", "age", "female", "edu_years", "cesd", "mmse")
)
}
\arguments{
\item{data}{input dataset}

\item{grouping_variable}{name of variable to group}

\item{grouping_values}{levels of the grouping variable}

\item{variables}{names of variables to obtain descriptive statistics (default: c('n', 'age', 'female', 'edu_years', 'cesd', 'mmse'))}
}
\value{
data.frame of descriptive statistics table
}
\description{
generate a descriptive statistics table with means and standard deviations overall and for each group
}
\concept{data_wrangling}
