% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duration.R
\name{get_duration}
\alias{get_duration}
\title{get_duration}
\usage{
get_duration(duration, message = "duration")
}
\arguments{
\item{duration}{difftime object}

\item{message}{message to display before the duration (default: duration)}
}
\value{

}
\description{
obtain duration in dd:hh:mm:sec format
}
\examples{
start_time <- Sys.time()
Sys.sleep(1)
get_duration(Sys.time() - start_time)
get_duration(as.difftime(1, units = 'secs'))
get_duration(as.difftime(1, units = 'mins'))
get_duration(as.difftime(1, units = 'hours'))
get_duration(as.difftime(1, units = 'days'))
get_duration(as.difftime(1.5251, units = 'days'), message = 'test')
}
\concept{helper}
