% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effect_size.R
\name{get_effect_size}
\alias{get_effect_size}
\title{get_effect_size}
\usage{
get_effect_size(model, a_thr = 0.05)
}
\arguments{
\item{model}{linear model object}

\item{a_thr}{alpha threshold for confidence interval (default = 0.05)}
}
\value{
data frame with effect size and confidence intervals
}
\description{
obtains effect size and confidence intervals for a given linear model
}
\examples{
model_fit <- lm(mpg ~ wt, data = mtcars)
get_effect_size(model_fit)
}
\concept{data_wrangling}
