% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_root_dir.R
\name{get_root_dir}
\alias{get_root_dir}
\title{get_root_dir}
\usage{
get_root_dir(directory)
}
\arguments{
\item{directory}{can be either kmk, kmr, or kenrod}
}
\value{
root_dir
}
\description{
get_root_dir
}
\examples{
get_root_dir()
}
\concept{helper}
