% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ses_from_path.R
\name{get_ses_from_path}
\alias{get_ses_from_path}
\title{get_ses_from_path}
\usage{
get_ses_from_path(path)
}
\arguments{
\item{path}{path}
}
\value{
ses
}
\description{
get_ses_from_path
}
\examples{
get_ses_from_path('/bids/sub-eep/ses-01/func/sub-eep_ses-01_bold.nii.gz')
}
\concept{helper}
