% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_from_airc_id.R
\name{get_sub_from_airc_id}
\alias{get_sub_from_airc_id}
\title{get_sub_from_airc_id}
\usage{
get_sub_from_airc_id(airc_id)
}
\arguments{
\item{airc_id}{AIRC ID}
}
\value{

}
\description{
get_sub_from_airc_id
}
\concept{helper}
