% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_from_path.R
\name{get_value_from_path}
\alias{get_value_from_path}
\title{get_value_from_path}
\usage{
get_value_from_path(path, key)
}
\arguments{
\item{path}{path}

\item{key}{key}
}
\value{
value
}
\description{
get_value_from_path
}
\examples{
get_value_from_path('/bids/sub-eep/ses-01/func/sub-eep_ses-01_bold.nii.gz', 'sub')
}
\concept{helper}
