% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanz.R
\name{meanz}
\alias{meanz}
\title{meanz}
\usage{
meanz(data)
}
\arguments{
\item{data}{data of variables}
}
\value{

}
\description{
z-score variables column-wise and then obtain the mean row-wise
}
\examples{
df <- carData::Salaries
yrs_z <- meanz(df[, c('yrs.since.phd', 'yrs.service')])
head(yrs_z)

summary(lm(salary ~ meanz(df[, c('yrs.since.phd', 'yrs.service')]), df))
}
\concept{data_wrangling}
