% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_age.R
\name{plot_by_age}
\alias{plot_by_age}
\title{plot_by_age}
\usage{
plot_by_age(
  data,
  age_var = "age",
  y_var = NULL,
  age_label = "Age",
  y_label = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{age_var}{}

\item{y_var}{}

\item{age_label}{}

\item{y_label}{}
}
\value{
plot with x-axis of age and y-axis of y_var
}
\description{
plot_by_age
}
\examples{
df_fake <- data.frame(age = sample(20:100, 100, TRUE),
                      outcome = sample(1:100/100, 100, TRUE))
plot_by_age(df_fake, y_var = 'outcome')
}
\concept{visualization}
