% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_change_plot.R
\name{plot_longitudinal_slopes_by_age}
\alias{plot_longitudinal_slopes_by_age}
\title{plot_longitudinal_slopes_by_age}
\usage{
plot_longitudinal_slopes_by_age(
  data_long,
  variable,
  age_var = "age",
  sub_var = "sub",
  wave_var = "wave"
)
}
\arguments{
\item{data_long}{}

\item{variable}{}

\item{age_var}{}

\item{sub_var}{}

\item{wave_var}{}
}
\value{

}
\description{
plot_longitudinal_slopes_by_age
}
