% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed_tidy.R
\name{renamed_tidy}
\alias{renamed_tidy}
\title{renamed_tidy}
\usage{
renamed_tidy(model)
}
\arguments{
\item{model}{input fitted model}
}
\value{
tidy with renamed columns
}
\description{
rename the columns to be more descriptive
}
\examples{
renamed_tidy(lm(salary ~ yrs.since.phd, carData::Salaries))
}
\concept{data_wrangling}
